+++ src/MoBi.Core/Services/InitialConditionsBuildingBlockExtendManager.cs
@@ -5,7 +5,6 @@ using MoBi.Core.Commands;
 using MoBi.Core.Domain.Extensions;
 using MoBi.Core.Domain.Model;
 using MoBi.Core.Domain.Services;
-using OSPSuite.Assets;
 using OSPSuite.Core.Domain;
 using OSPSuite.Core.Domain.Builder;
 using OSPSuite.Core.Domain.Services;
@@ -22,17 +21,15 @@ public interface IInitialConditionsBuildingBlockExtendManager : IExtendPathAndVa
 public class InitialConditionsBuildingBlockExtendManager : ExtendPathAndValuesManager<InitialCondition>, IInitialConditionsBuildingBlockExtendManager
 {
    private readonly IInitialConditionsCreator _initialConditionsCreator;
-   private readonly IDimensionFactory _dimensionFactory;
 
    public InitialConditionsBuildingBlockExtendManager(
       IInitialConditionsCreator initialConditionsCreator,
       IMoBiFormulaTask moBiFormulaTask,
       IObjectTypeResolver objectTypeResolver,
       IMoBiContext moBiContext,
-      IDimensionFactory dimensionFactory) : base(moBiFormulaTask, objectTypeResolver, moBiContext)
+      IDimensionFactory dimensionFactory) : base(moBiFormulaTask, objectTypeResolver, moBiContext, dimensionFactory)
    {
       _initialConditionsCreator = initialConditionsCreator;
-      _dimensionFactory = dimensionFactory;
    }
 
    private void updateDefaultIsPresentToFalseForSpecificExtendedValues(IReadOnlyList<InitialCondition> allInitialConditions, IReadOnlyList<InitialCondition> templateValues)
@@ -58,7 +55,10 @@ public class InitialConditionsBuildingBlockExtendManager : ExtendPathAndValuesMa
    {
       var pathAndValueEntity = buildingBlock[properties.ObjectPath];
       if (pathAndValueEntity != null)
-         return updateValueAndDimension(buildingBlock, pathAndValueEntity, properties);
+      {
+         var updateValueCommand = new UpdateInitialConditionInBuildingBlockCommand(buildingBlock, properties.ObjectPath, properties.ValueInBaseUnit, properties.IsPresent, properties.ScaleDivisor, properties.NegativeAllowed);
+         return UpdateValueAndDimension(updateValueCommand, buildingBlock, pathAndValueEntity, properties.DimensionName);
+      }
 
       var moleculeName = properties.ObjectPath.Last();
       var containerPath = properties.ObjectPath.Clone<ObjectPath>();
@@ -78,24 +78,5 @@ public class InitialConditionsBuildingBlockExtendManager : ExtendPathAndValuesMa
       );
 
       return GenerateAddCommand(buildingBlock, initialCondition);
-   }
-
-   private IMoBiCommand updateValueAndDimension(InitialConditionsBuildingBlock buildingBlock, InitialCondition pathAndValueEntity, InitialConditionPropertiesForMerge properties)
-   {
-      var updateValueCommand = new UpdateInitialConditionInBuildingBlockCommand(buildingBlock, properties.ObjectPath, properties.ValueInBaseUnit, properties.IsPresent, properties.ScaleDivisor, properties.NegativeAllowed);
-
-      var newDimension = _dimensionFactory.Dimension(properties.DimensionName);
-      if (Equals(pathAndValueEntity.Dimension, newDimension))
-         return updateValueCommand;
-
-      var macroCommand = new MoBiMacroCommand
-      {
-         CommandType = AppConstants.Commands.UpdateCommand,
-         ObjectType = ObjectTypes.InitialCondition
-      };
-
-      macroCommand.Add(updateValueCommand);
-      macroCommand.Add(new UpdateDimensionInPathAndValueEntityCommand<InitialCondition>(pathAndValueEntity, newDimension, newDimension.DefaultUnit, buildingBlock));
-      return macroCommand;
    }
 }
