+++ src/MoBi.Core/Services/ParameterValueBuildingBlockExtendManager.cs
@@ -1,9 +1,7 @@
 ﻿using System.Collections.Generic;
-using MoBi.Assets;
 using MoBi.Core.Commands;
 using MoBi.Core.Domain.Model;
 using MoBi.Core.Domain.Services;
-using OSPSuite.Assets;
 using OSPSuite.Core.Domain;
 using OSPSuite.Core.Domain.Builder;
 using OSPSuite.Core.Domain.Services;
@@ -19,12 +17,10 @@ public interface IParameterValueBuildingBlockExtendManager : IExtendPathAndValue
 public class ParameterValueBuildingBlockExtendManager : ExtendPathAndValuesManager<ParameterValue>, IParameterValueBuildingBlockExtendManager
 {
    private readonly IParameterValuesCreator _parameterValuesCreator;
-   private readonly IDimensionFactory _dimensionFactory;
 
-   public ParameterValueBuildingBlockExtendManager(IParameterValuesCreator parameterValuesCreator, IMoBiFormulaTask moBiFormulaTask, IObjectTypeResolver objectTypeResolver, IMoBiContext moBiContext, IDimensionFactory dimensionFactory) : base(moBiFormulaTask, objectTypeResolver, moBiContext)
+   public ParameterValueBuildingBlockExtendManager(IParameterValuesCreator parameterValuesCreator, IMoBiFormulaTask moBiFormulaTask, IObjectTypeResolver objectTypeResolver, IMoBiContext moBiContext, IDimensionFactory dimensionFactory) : base(moBiFormulaTask, objectTypeResolver, moBiContext, dimensionFactory)
    {
       _parameterValuesCreator = parameterValuesCreator;
-      _dimensionFactory = dimensionFactory;
    }
 
    protected override IReadOnlyList<ParameterValue> CreatePathAndValueEntitiesBasedOnUsedTemplates(SpatialStructure spatialStructure, IReadOnlyList<MoleculeBuilder> molecules, ILookupBuildingBlock<ParameterValue> buildingBlock) =>
@@ -38,7 +34,10 @@ public class ParameterValueBuildingBlockExtendManager : ExtendPathAndValuesManag
    {
       var pathAndValueEntity = buildingBlock[objectPath];
       if (pathAndValueEntity != null)
-         return updateValueAndDimension(buildingBlock, pathAndValueEntity, objectPath, valueInBaseUnit, dimensionName);
+      {
+         var updateValueCommand = new UpdateParameterValueInBuildingBlockCommand(buildingBlock, objectPath, valueInBaseUnit);
+         return UpdateValueAndDimension(updateValueCommand, buildingBlock, pathAndValueEntity, dimensionName);
+      }
 
       var dimension = _dimensionFactory.Dimension(dimensionName);
 
@@ -51,24 +50,5 @@ public class ParameterValueBuildingBlockExtendManager : ExtendPathAndValuesManag
       );
 
       return GenerateAddCommand(buildingBlock, parameterValue);
-   }
-
-   private IMoBiCommand updateValueAndDimension(ParameterValuesBuildingBlock buildingBlock, ParameterValue pathAndValueEntity, ObjectPath objectPath, double valueInBaseUnit, string dimensionName)
-   {
-      var updateValueCommand = new UpdateParameterValueInBuildingBlockCommand(buildingBlock, objectPath, valueInBaseUnit);
-
-      var newDimension = _dimensionFactory.Dimension(dimensionName);
-      if (Equals(pathAndValueEntity.Dimension, newDimension))
-         return updateValueCommand;
-
-      var macroCommand = new MoBiMacroCommand
-      {
-         CommandType = AppConstants.Commands.UpdateCommand,
-         ObjectType = ObjectTypes.ParameterValue
-      };
-
-      macroCommand.Add(updateValueCommand);
-      macroCommand.Add(new UpdateDimensionInPathAndValueEntityCommand<ParameterValue>(pathAndValueEntity, newDimension, newDimension.DefaultUnit, buildingBlock));
-      return macroCommand;
    }
 }
